🛠️ refactor: 재사용 컴포넌트를 위한 common 폴더 생성 및 FormInput 이동
